+++ src/generate/generic_atomic.rs
@@ -1,5 +1,5 @@
 mod atomic {
-    use portable_atomic::{AtomicU16, AtomicU8, Ordering};
+    use portable_atomic::Ordering;
 
     pub trait AtomicOperations {
         unsafe fn atomic_or(ptr: *mut Self, val: Self);
@@ -11,21 +11,30 @@ mod atomic {
         ($U:ty, $Atomic:ty) => {
             impl AtomicOperations for $U {
                 unsafe fn atomic_or(ptr: *mut Self, val: Self) {
-                    (*(ptr as *const $Atomic)).fetch_or(val, Ordering::SeqCst);
+                    (*(ptr as *const $Atomic)).or(val, Ordering::SeqCst);
                 }
 
                 unsafe fn atomic_and(ptr: *mut Self, val: Self) {
-                    (*(ptr as *const $Atomic)).fetch_and(val, Ordering::SeqCst);
+                    (*(ptr as *const $Atomic)).and(val, Ordering::SeqCst);
                 }
 
                 unsafe fn atomic_xor(ptr: *mut Self, val: Self) {
-                    (*(ptr as *const $Atomic)).fetch_xor(val, Ordering::SeqCst);
+                    (*(ptr as *const $Atomic)).xor(val, Ordering::SeqCst);
                 }
             }
         };
     }
-    impl_atomics!(u8, AtomicU8);
-    impl_atomics!(u16, AtomicU16);
+
+    impl_atomics!(u8, portable_atomic::AtomicU8);
+    impl_atomics!(u16, portable_atomic::AtomicU16);
+
+    // Exclude 16-bit archs from 32-bit atomics
+    #[cfg(not(target_pointer_width = "16"))]
+    impl_atomics!(u32, portable_atomic::AtomicU32);
+
+    // Enable 64-bit atomics for 64-bit RISCV
+    #[cfg(any(target_pointer_width = "64", target_has_atomic = "64"))]
+    impl_atomics!(u64, portable_atomic::AtomicU64);
 }
 use atomic::AtomicOperations;